% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_rq.R
\name{calculate_rq}
\alias{calculate_rq}
\title{Calculate RQ}
\usage{
calculate_rq(data, tar, ref)
}
\arguments{
\item{data}{a tibble provided by calculate_n0().}

\item{tar}{a character vector indicating the names of the target amplicons.}

\item{ref}{a character vector indicating the names of the reference amplicons.}
}
\value{
A tibble containing the RQ values for each sample. RQ values are computed as the ratio target mean_n0 / reference mean_n0 and stored in the rq_val variable.
}
\description{
calculate_rq() computes quantification of target amplicons relative to reference amplicons.
}
\examples{

}
