% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddct_compute.R
\name{compute_ddct}
\alias{compute_ddct}
\title{Compute delta delta Ct}
\usage{
compute_ddct(data, hkg, cal)
}
\arguments{
\item{data}{a dataframe generated by format_sop().}

\item{hkg}{a character vector indicating the name(s) of the reference gene(s) to use for normalisation.}

\item{cal}{a character vector indicating the name of the reference sample to use for normalisation.}
}
\value{
A list storing the mean Ct values in "raw.data", and the relative quantification results in "rel.quant".
}
\description{
compute_ddct() calculates delta delta Ct from Ct values stored in a dataframe generated by format_sop().
}
\examples{

}
