% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_fluostar.omega_data.R
\name{format_fluostar.omega_data}
\alias{format_fluostar.omega_data}
\title{Format fluorescence data from FLUOstar Omega plate reader}
\usage{
format_fluostar.omega_data(data_file, infos_samples_file, extend = FALSE)
}
\arguments{
\item{data_file}{blabla}

\item{infos_samples_file}{blabla}

\item{extend}{blabla}
}
\value{
blabla
}
\description{
blabla
}
\examples{

}
