% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_ms_data.R
\name{format_ms_data}
\alias{format_ms_data}
\title{Tidy MS report}
\usage{
format_ms_data(filename)
}
\arguments{
\item{filename}{a txt file derived from a mht MS data file. In Microsoft Word: Open the mht file, then "Save As..." > "File Format: Plain Text" > "Text encoding: Unicode UTF-8".}
}
\value{
A tibble summarising the Mascot search results.
}
\description{
format_ms_data() tidies Mascot search results provided by the Cambridge Centre for Proteomics (CCP) for further processing in R.
It takes a MS report saved as a txt file as input and returns a tibble summarising the results.
}
\examples{

}
