% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_sop.R
\name{format_sop}
\alias{format_sop}
\title{Format qPCR results from Step One Plus software}
\usage{
format_sop(file)
}
\arguments{
\item{file}{a string specifying the path to the .csv file containing qPCR results.}
}
\value{
A dataframe summarising results in 3 columns: sample.id, tar.nm, ct.
}
\description{
format_sop() tidies the "Results" sheet provided by StepOne Software v2.3, and makes it ready for processing by ddct.compute().
}
\examples{

}
