% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_amplicons.R
\name{gather_amplicons}
\alias{gather_amplicons}
\title{Gather identical amplicons from distinct CFX output files}
\usage{
gather_amplicons(path, datafiles, max_out = FALSE)
}
\arguments{
\item{path}{a string specifying the path to the directory containing LinRegPCR output directories.}

\item{datafiles}{a character vector containing the names of LinRegPCR output directories. Each LinRegPCR output directory must be named after its corresponding CFX data file and must contain the "Quantification Cq Results" and "Quantification Amplification Results" files provided by CFX software. These files must be exported as csv from CFX software and their names contain, as a prefix, the name of the corresponding CFX data file.}

\item{max_out}{a logical, default is FALSE. If TRUE, all amplicons are gathered on a single set of virtual CFX output. So, LinRegPCR software can be fed with a single pair of virtual infos_plate and fluo_data files.}
}
\value{
A tibble containing virtual infos_plate and fluo_data variables, either for each amplicon (max_out = FALSE) or for all amplicons gathered together (max_out = TRUE).
}
\description{
gather_amplicons() aims at creating virtual CFX output as if the amplification of a given cDNA, performed on distinct qPCR plates, had been performed on a single plate.

It creates a single set of virtual CFX output for each amplicon spread across distinct original CFX output files. Virtual infos_plate and fluo_data variables are created for each amplicon and can then be exported as virtual CFX output files to feed LinRegPCR software.
}
\examples{

}
