% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_avp.R
\name{get_avp}
\alias{get_avp}
\title{Add added-variable plots}
\usage{
get_avp(tibble)
}
\arguments{
\item{tibble}{a tibble produced by choose_on_aic() or get_sign_mod().}
}
\value{
A tibble in which the added-variable plots for each best model are stored in a variable called avp.
}
\description{
get_avp() draws the added-variable plots for the "best models" determined by choose_on_aic().
}
\examples{

}
