% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_complex_anova_beta.R
\name{get_complex_anova_beta}
\alias{get_complex_anova_beta}
\title{Blabla}
\usage{
get_complex_anova_beta(
  tibble,
  grp = NULL,
  lhs,
  rhs,
  base.size = 14,
  legend = "none",
  spe.contr = FALSE,
  specs = NA_character_,
  by = NA_character_,
  adjust.emm.comp = "tukey",
  block = NA_character_,
  jumps = 0
)
}
\arguments{
\item{tibble}{blabla}

\item{grp}{blabla}

\item{lhs}{blabla}

\item{rhs}{blabla}

\item{base.size}{blabla}

\item{legend}{blabla}

\item{spe.contr}{blabla}

\item{specs}{blabla}

\item{by}{blabla}

\item{adjust.emm.comp}{blabla}

\item{block}{blabla}

\item{jumps}{blabla}
}
\value{
blabla
}
\description{
Blabla
}
\examples{

}
