% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inrange_flowdata.R
\name{get_inrange_flowdata}
\alias{get_inrange_flowdata}
\title{Get in range flow cytometry data}
\usage{
get_inrange_flowdata(fs)
}
\arguments{
\item{fs}{a flowSet to tidy.}
}
\value{
a tidied flowSet.
}
\description{
get_inrange_flowdata() "tidies" each flowFrame of a flowSet by removing all events for which the recorded value of one or more parameter is out of range.
}
\examples{

}
