% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameter_voltage.R
\name{get_parameter_voltage}
\alias{get_parameter_voltage}
\title{Get voltage values of flow cytometry parameters}
\usage{
get_parameter_voltage(fs, check_identical = FALSE)
}
\arguments{
\item{fs}{a flowSet to analyse.}

\item{check_identical}{a logical indicating if the equality of voltage values should be checked.}
}
\value{
a data frame summarizing the voltage values of parameters included in the flowSet. If check_identical = TRUE, get_parameter_voltage() returns a list composed of the voltage_table and the voltage_check.
}
\description{
get_parameter_voltage() extracts the voltage value of parameters in each flowFrame of a flowSet. If requested, it also checks that all voltage values of a parameter are identical across the flowFrames.
}
\examples{

}
