% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shared_keywords.R
\name{get_shared_keywords}
\alias{get_shared_keywords}
\title{Get shared keyword values in a flowSet/GatingSet}
\usage{
get_shared_keywords(flow.obj, invert = FALSE)
}
\arguments{
\item{flow.obj}{a flowSet or GatingSet to analyse.}

\item{invert}{if TRUE, return the keyword values not shared.}
}
\value{
a data frame summarizing the keyword values shared (or not shared) between the elements of the flowSet/GatingSet.
}
\description{
get_shared_keywords() summarizes the keyword values shared across elements of a flowSet or GatingSet. If requested, it can return the keyword values which are not shared instead.
}
\examples{

}
