% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_voltage_settings.R
\name{get_voltage_settings}
\alias{get_voltage_settings}
\title{Get voltage settings of flow cytometry parameters}
\usage{
get_voltage_settings(fs, check_identical = FALSE)
}
\arguments{
\item{fs}{a flowSet to analyse.}

\item{check_identical}{a logical indicating if the equality of voltage settings should be checked.}
}
\value{
a data frame summarizing the voltage settings of parameters included in the flowSet. If check_identical = TRUE, get_voltage_settings() returns a list composed of the voltage_table and the voltage_check.
}
\description{
get_voltage_settings() extracts the voltage settings of parameters in each flowFrame of a flowSet. If requested, it also checks that all voltage settings of a parameter are identical across the flowFrames.
}
\examples{

}
