% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cell_profiler_data.R
\name{import_cell_profiler_data}
\alias{import_cell_profiler_data}
\title{Import CellProfiler data}
\usage{
import_cell_profiler_data(
  RelateObjects.data.files.list,
  IdentifyPrimaryObjects.data.files.list,
  prefix,
  main.object.nm = "neuron",
  main.object.structure.nm = "cell_body",
  metadata = FALSE
)
}
\arguments{
\item{RelateObjects.data.files.list}{a string specifying the name of the .txt file which lists the names of all RelateObjects data files to be analysed. In the .txt list file, the RelateObjects file names must match the regex "RelateObjects_.*.csv" and be separated by a new line.}

\item{IdentifyPrimaryObjects.data.files.list}{a string specifying the name of the .txt file which lists the names of all IdentifyPrimaryObjects data files to be analysed. In the .txt list file, the IdentifyPrimaryObjects file names must match the regex "IdentifyPrimaryObjects_.*.csv" and be separated by a new line.}

\item{prefix}{a string specifying the path to the directory containing the .txt list files & .csv data files.}

\item{main.object.nm}{a string indicating the name to be given to the main object defined by the main IdentifyPrimaryObjects module. The main object is the one (e.g.: neuron) within which other objects (e.g.: pSyn structures) are defined by additional IdentifyPrimaryObjects modules. Default is "neuron".}

\item{main.object.structure.nm}{a string indicating the name used in the main IdentifyPrimaryObjects module to define the main object in the CellProfiler pipeline. Default is "cell_body".}

\item{metadata}{a logical indicating if scaling-related metadata should be imported together with the CellProfiler data. If TRUE, the metadata should be provided as .xml files stored in a directory named "metadata" and located in the same directory as the .txt list files & .csv data files. Default is FALSE.}
}
\value{
a tibble containing the tidied CellProfiler data (and possibly the associated metadata). Each row corresponds to a main object, columns store information related to accessory objects identified within the main object.
}
\description{
import_cell_profiler_data() imports and tidy data generated by CellProfiler software.
}
\details{
Important note: the function requires all input files (.txt list files, .csv data files and metadata directory containing .xml metadata files) to be located in the same directory.
}
\examples{

}
