% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore_amp_names.R
\name{restore_amp_names}
\alias{restore_amp_names}
\title{Restore amplicons names in LinRegPCR output}
\usage{
restore_amp_names(path, filenames)
}
\arguments{
\item{path}{a string specifying the path to the directory containing LinRegPCR output files.}

\item{filenames}{a character vector containing the names of xlsx files produced by LinRegPCR, with the xlsx extension included. The files must be of xlsx type and their name must match the following regex pattern: "^.*_amplicon.name.xlsx$".}
}
\value{
For each xlsx file provided as input, a csv file in which the amplicon name has been restored. The csv files are written in the directory indicated by the path argument.
}
\description{
restore_amp_names() is useful when LinRegPCR output files contain a single amplicon, in which case the amplicon name is replaced by "not_named".

The function takes xlsx files produced by LinRegPCR as inputs and returns csv files in which the amplicon names have been restored based on files names.
}
\examples{

}
