% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cell_profiler_data.R
\name{import_cell_profiler_data}
\alias{import_cell_profiler_data}
\title{Import CellProfiler data}
\usage{
import_cell_profiler_data(
  RelateObjects.data.files.list,
  IdentifyPrimaryObjects.data.files.list,
  prefix,
  main.object.nm = "neuron",
  main.object.structure.nm = "cell_body",
  metadata = FALSE
)
}
\arguments{
\item{RelateObjects.data.files.list}{blabla}

\item{IdentifyPrimaryObjects.data.files.list}{blabla}

\item{prefix}{blabla}

\item{main.object.nm}{blabla}

\item{main.object.structure.nm}{blabla}

\item{metadata}{blabla}
}
\value{
blabla
}
\description{
blabla
}
\examples{

}
